Update start_local_popit_api.bash use current PopIt
The start_local_popit_api.bash script is designed to make it easy to set
up a local version of PopIt for testing against; this is still useful
for re-recording the fixtures (see the commit message for 967f842529
and https://gist.github.com/mhl/848f7980d256ca38ff603375be4f264e ).

However, this script installed a very old version of PopIt - this
commit updates it to use the most recent version. (A key problem with
the old version of this script is that it sets up a version of PopIt
without the export.json endpoints.)