Add a page listing missing contacts and appealing for email addresses
We're not sure if this will work, but several people have suggeted
that if there's no contact email address for a person, then it would
be worth appealing to users of the site to find out email addresses
for people.

How they feed this back to the instance owner is currently an
incomplete story.  The owner's email address should not
be exposed on the site, so we can't just include that.  For the
iran subdomain we have a particular support email address that
we're suggesting in the overridden template, but for the generic
case I suggest we create a partial template that can be included
anywhere which provides a simple contact form for sending a
message to the instance's owner.  This might need captcha protection,
etc. but a basic version would be quick to implement.