Django 1.8: get Celery working with the new Django version
We switch to using Celery in the manner recommended in Celery 3.1 and
later. We wouldn't need django-celery any more, except that we're using
it provides a Django ORM-based results backend for Celery.

We upgrade django-celery to avoid some deprecation warnings.

The celery worker and beat are no longer invoked using manage.py but
with the celery script.