Make the base site's direction RTL based on the language
When going to the base site (i.e. without an instance subdomain)
in a right-to-left language, the page was missing dir="rtl", which
meant that direction-neutral characters like '.' were being put at
the right hand side of sentences.  This changes the template logic
to add dir="html" on the <html> element if the i18n framework
says that this is a right-to-left language.