Add a template loader that looks for subdomain-specific templates
In combination with the middleware that stores the request's subdomain
in thread local storage, this template loader acts very like the
built-in filesystem template loader, but inserts
'subdomains/<subdomain>' before the template name in the path that's
looked up. This is so that we can override the templates used for
particular domains.