Allow filtering messages in API by Popolo URI
This adds a 'person__popolo_uri' parameter to the API which works in a
similar way to the existing 'person__popit_id' filter, but looks at the
identifiers with a 'popolo_uri' scheme.

The hardest part of this change was writing the test for it. I'm still
not totally happy with how it works, but it at least works as a
regression test for now. Suggestions welcome!