Pin psycopg2 to 2.6.2 in requirements-mysociety.txt
Pinning to this version prevents the following warning from being
produced:

    UserWarning: The psycopg2 wheel package will be renamed from release
    2.8; in order to keep installing from binary please use "pip install
    psycopg2-binary" instead. For details see:
    <http://initd.org/psycopg/docs/install.html#binary-install-from-pypi>.

This in itself is harmless, but the warning was being printed from the
handleemail managements command, which causes exim to send an email from
its "Mail Delivery System" address because it has the `return_output`
option enabled.

A better long term fix would be to disable `return_output` in the exim
config for the WiP hostname, but in the meantime lets just fix this
specific warning.